Add mirrored switch holders
