+++ klavgen/__init__.py
@@ -15,8 +15,6 @@ from . import (
     renderer_screw_hole,
     renderer_switch_holder,
     renderer_switch_holder_choc,
-    renderer_switch_holder_choc_bottom,
-    renderer_switch_holder_choc_top,
     renderer_trrs_jack,
     renderer_usbc_jack,
     rendering,
@@ -26,8 +24,6 @@ importlib.reload(rendering)
 importlib.reload(classes)
 importlib.reload(config)
 importlib.reload(renderer_switch_holder_choc)
-importlib.reload(renderer_switch_holder_choc_bottom)
-importlib.reload(renderer_switch_holder_choc_top)
 importlib.reload(constants)
 importlib.reload(keyboard)
 importlib.reload(kle)
