+++ klavgen/renderer_case.py
@@ -34,6 +34,8 @@ class RenderCaseResult:
     switch_holes: Optional[cq.Workplane] = None
     fused_switch_holders: Optional[cq.Workplane] = None
     fused_switch_holder_clearances: Optional[cq.Workplane] = None
+    fused_switch_holders_mirrored: Optional[cq.Workplane] = None
+    fused_switch_holder_clearances_mirrored: Optional[cq.Workplane] = None
     inner_clearances: Optional[cq.Workplane] = None
     patches: Optional[cq.Workplane] = None
     cuts: Optional[cq.Workplane] = None
@@ -55,6 +57,7 @@ class RenderCaseResult:
     bottom_before_fillet: Optional[cq.Workplane] = None
     bottom: Optional[cq.Workplane] = None
     switch_holders: Optional[cq.Workplane] = None
+    switch_holders_mirrored: Optional[cq.Workplane] = None
     debug: Optional[cq.Workplane] = None
     standard_components: Optional[cq.Workplane] = None
     components: Optional[Dict[str, Dict[str, List[cq.Workplane]]]] = None
@@ -69,6 +72,7 @@ def render_case(
     case_extras: Optional[List[cq.Workplane]] = None,
     palm_rests: Optional[List[PalmRest]] = None,
     texts: Optional[List[Text]] = None,
+    mirror_switch_holders: bool = False,
     debug: bool = False,
     render_standard_components: bool = False,
     result: Optional[RenderCaseResult] = None,
@@ -84,6 +88,8 @@ def render_case(
     :param case_extras: A list of CadQuery objects to be added to the case. Can be used for custom outlines. Optional.
     :param palm_rests: A list of PalmRest objects that define palm rests (overlapping with keys is fine). Optional.
     :param texts: A list of Text objects with text to be drawn to either the top or palm rests. Optional.
+    :param mirror_switch_holders: Whether to also produce a mirrored version of the Choc switch holders. Needs to be
+                                  use on split keyboards where both sides are the same.
     :param debug: A boolean defining whether to run in debug mode which skips the finicky shell step (that produces
                   the empty internal volume of the case.
     :param render_standard_components: A boolean defining whether to render all the holders and connectors in the
@@ -175,11 +181,23 @@ def render_case(
 
     result.switch_holes = union_list([rk.switch_hole for rk in rendered_keys])
 
+    # Fused Choc switch holders
     result.fused_switch_holders = union_list([rk.fused_switch_holder for rk in rendered_keys])
-
     result.fused_switch_holder_clearances = union_list(
         [rk.fused_switch_holder_clearance for rk in rendered_keys]
     )
+
+    # Mirrored fused Choc switch holders for split keyboards
+    result.fused_switch_holders_mirrored = None
+    result.fused_switch_holder_clearances_mirrored = None
+    if mirror_switch_holders:
+        result.fused_switch_holders_mirrored = union_list(
+            [rk.fused_switch_holder_mirrored for rk in rendered_keys]
+        )
+
+        result.fused_switch_holder_clearances_mirrored = union_list(
+            [rk.fused_switch_holder_clearance_mirrored for rk in rendered_keys]
+        )
 
     result.debug = union_list([rendered_key.debug for rendered_key in rendered_keys])
 
@@ -540,6 +558,7 @@ def render_case(
 
     result.debug = union_list(result.debug, stage_rendered_components[RenderingPipelineStage.DEBUG])
 
+    # Fused Choc switch holder
     result.switch_holders = None
     if result.fused_switch_holders:
         result.switch_holders = result.fused_switch_holders
@@ -551,6 +570,28 @@ def render_case(
             result.switch_holders = result.switch_holders.cut(result.inner_clearances)
 
         result.switch_holders = result.switch_holders.intersect(result.inner_volume_after_fillet)
+
+    # Mirrored fused Choc switch holders
+    result.switch_holders_mirrored = None
+    if result.fused_switch_holders_mirrored:
+        result.switch_holders_mirrored = result.fused_switch_holders_mirrored
+
+        if result.fused_switch_holder_clearances_mirrored:
+            result.switch_holders_mirrored = result.switch_holders_mirrored.cut(
+                result.fused_switch_holder_clearances_mirrored
+            )
+
+        if result.inner_clearances:
+            result.switch_holders_mirrored = result.switch_holders_mirrored.cut(
+                result.inner_clearances
+            )
+
+        result.switch_holders_mirrored = result.switch_holders_mirrored.intersect(
+            result.inner_volume_after_fillet
+        )
+
+        # Flip so the mirrored switch holders are in the right orientation
+        result.switch_holders_mirrored = result.switch_holders_mirrored.mirror(mirrorPlane="YZ")
 
     if result.shell_cut:
         result.bottom = result.bottom.cut(result.shell_cut)
@@ -684,8 +725,16 @@ def export_case_to_stl(result: RenderCaseResult):
                 cq.exporters.export(palm_rest, f"palm_rest_{index}.stl")
 
     if result.switch_holders:
+        # Flip around Y so it's in printing orientation
         switch_holders_oriented = result.switch_holders.rotate((0, 0, 0), (0, 1, 0), 180)
         cq.exporters.export(switch_holders_oriented, "switch_holders.stl")
+
+    if result.switch_holders_mirrored:
+        # Flip around Y so it's in printing orientation
+        switch_holders_mirrored_oriented = result.switch_holders_mirrored.rotate(
+            (0, 0, 0), (0, 1, 0), 180
+        )
+        cq.exporters.export(switch_holders_mirrored_oriented, "switch_holders_mirrored.stl")
 
 
 def export_case_to_step(result: RenderCaseResult):
